etl/requirements.txt
backoff==2.2.1
elasticsearch==8.6.2
psycopg2==2.9.5
psycopg2_binary==2.9.5
python-dotenv==1.0.0
pyyaml==6.0